fix(detector): 🐛 修复所有优先级 Bug（P0→P1→P2），新增 39 项专项测试
P0 修复:
- fusion_engine.py: 候选合并 bbox 改为外包矩形，保留完整空间证据
- roi_locator.py: 添加 NaN/Inf 显式检查，防止 ONNX 输出异常静默传播

P1 修复:
- preprocessor.py: 透视翘曲 float64→float32，内存减半 (768→384 MB)
- ecc_registration.py: 暴力搜索新增半径硬上限(10px)和迭代硬上限，防 O(R²) 爆炸

P2 修复:
- quality_gate.py: 机器人位姿容差由硬编码 1e-4 改为 recipe.quality.max_pose_delta 可配置
- feature_builder.py: 通道别名表添加弃用说明，引导使用 light:ID 规范语法

新增测试 (39 tests):
- test_ecc_registration.py: ECC 配准/回退/安全约束 (9 tests)
- test_roi_locator_mask.py: NaN 防御/上采样/腐蚀 (6 tests)
- test_fusion_engine.py: 外包矩形合并边界 (6 new tests)
- test_rule_engine_thresholds.py: 质量失败/错误/长宽比 (12 new tests)
- test_yolo_decode_seg.py: YOLO 解码/Seg 候选 (6 tests)

验证: 274/274 tests passed, protocol + model assets OK

+++ python_detector/pipeline/preprocessor.py
@@ -320,7 +320,7 @@ class Preprocessor:
         output_height: int,
         roi_name: str,
     ) -> bytearray:
-        grid_y, grid_x = np.indices((output_height, output_width), dtype=np.float64)
+        grid_y, grid_x = np.indices((output_height, output_width), dtype=np.float32)
         denom = transform[6] * grid_x + transform[7] * grid_y + transform[8]
         if np.any(np.abs(denom) < 1e-9):
             raise ValueError(f"{frame.camera_id}/{frame.light_id}/{roi_name}: ROI 透视变换无效")
@@ -331,7 +331,7 @@ class Preprocessor:
 
     def _sample_bilinear_grid(self, frame: LightFrame, x: np.ndarray, y: np.ndarray) -> np.ndarray:
         source = np.frombuffer(frame.image, dtype=np.uint8, count=frame.stride_bytes * frame.height)
-        source = source.reshape(frame.height, frame.stride_bytes)[:, : frame.width].astype(np.float64, copy=False)
+        source = source.reshape(frame.height, frame.stride_bytes)[:, : frame.width].astype(np.float32, copy=False)
         clipped_x = np.clip(x, 0.0, float(frame.width - 1))
         clipped_y = np.clip(y, 0.0, float(frame.height - 1))
         x0 = clipped_x.astype(np.int64)
